Warp back to 5578.  Sorry for the lost changes, please re-apply.  Reason: 5579 is a complete disaster.
Not only does it change tons of files to switch to a new and non-working (it doesn't parse my ini files, at least) ini parser, it also reshuffles a lot of code and removes a plugin. The latter part is fine, but doing these two major switches in one revision, one of which is broken, is completely unacceptable. I said to merge tiny changes, not massive reworkings.

git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@5589 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DiscIO/Src/FileMonitor.cpp
@@ -142,7 +142,7 @@ void FindFilename(u64 offset)
 	if (!fname || (strlen(fname) == 512))
 		return;
 	
-	CheckFile(fname, (int)pFileSystem->GetFileSize(fname));
+	CheckFile(fname, pFileSystem->GetFileSize(fname));
 }
 
 void Close()
